+++ src/core/index.ts
@@ -2,6 +2,7 @@
 export * from './types/index.js';
 export * from './interfaces/index.js';
 export * from './redis-service.js';
+export * from './job-broker.js';
 export * from './message-handler.js';
 export * from './connection-manager.js';
 export * from './utils/index.js';