+++ src/core/interfaces/index.ts
@@ -5,3 +5,4 @@ export * from './redis-service.js';
 export * from './message-handler.js';
 export * from './connection-manager.js';
 export * from './message-router.js';
+export * from './job-broker.js';